refactor: Remove old configmapsecret components, utils, etc.; configmapsecret services moved to new folder
