--- src/Pages/Shared/ConfigMapSecretOld/ConfigMapSecret.scss
@@ -1,112 +1,0 @@
-/*
- * Copyright (c) 2024. Devtron Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.form-row__select-external-type {
-    justify-content: flex-start;
-
-    .edit {
-        margin-left: auto;
-    }
-
-    .popup-button {
-        width: 50%;
-    }
-
-    .select-button {
-        height: 40px;
-        padding-left: 10px;
-    }
-}
-
-.cm-secret-container {
-    &.with-comment-drawer {
-        display: grid;
-        grid-template-columns: 1fr 300px;
-        position: absolute;
-        left: 56px;
-        top: 77px;
-        right: 0;
-        height: calc(100% - 77px);
-        animation: comments-view-opening;
-        animation-duration: 200ms;
-        animation-timing-function: linear;
-        animation-fill-mode: forwards;
-
-        & .crud-btn {
-            max-width: calc(100% - 300px);
-        }
-    }
-
-    &.with-crud-btn {
-        >.main-content {
-            height: calc(100vh - 142px);
-        }
-    }
-
-    @keyframes comments-view-opening {
-        0% {
-            grid-template-columns: 1fr 0;
-        }
-
-        100% {
-            grid-template-columns: 1fr 300px;
-        }
-    }
-
-    .tippy-over {
-        .tippy-box {
-            top: 42px;
-        }
-    }
-
-    .approval-config-disabled {
-        opacity: 0.5;
-        cursor: not-allowed;
-        background: var(--B500);
-        color: var(--white);
-        border: 1px solid transparent;
-    }
-}
-
-.variables-widget-position-cmcs {
-    position: absolute;
-    bottom: 100px;
-    right: 160px;
-}
-
-.validation-error-block {
-    padding: 8px 16px;
-    background: var(--R100);
-    display: grid;
-    grid-template-columns: 16px 1fr;
-    grid-column-gap: 8px;
-    align-content: initial;
-    color: #f32e2e;
-    font-size: 12px;
-
-    svg {
-        height: 16px;
-        width: 16px;
-    }
-
-    white-space: pre;
-}
-
-.sub-path-checkbox {
-    .form__checkbox-container {
-        align-self: flex-start;
-    }
-}
